BGDIDIC-1783: rework lfp2/hfp2
* new datasource WFS from geodienste.ch
* slightly changed data structure
* new target table in stopo_master
