+++ chsdi/models/vector/kogis.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from sqlalchemy import Column, Integer
-from sqlalchemy.types import Numeric, Unicode
+from sqlalchemy.types import Unicode
 
 from chsdi.models import register, bases
 from chsdi.models.vector import Vector, Geometry2D
@@ -21,51 +21,3 @@ class Agnes(Base, Vector):
     the_geom = Column(Geometry2D)
 
 register('ch.swisstopo.fixpunkte-agnes', Agnes)
-
-
-class FixpunkteLfp2(Base, Vector):
-    __tablename__ = 'punkt_lage_lfp2'
-    __table_args__ = ({'schema': 'fpds', 'autoload': False})
-    __template__ = 'templates/htmlpopup/fixpunkte.mako'
-    __bodId__ = 'ch.swisstopo.fixpunkte-lfp2'
-    __label__ = 'id'
-    id = Column('pointid', Unicode, primary_key=True)
-    nbident = Column('nbident', Unicode)
-    punktname = Column('punktname', Unicode)
-    status = Column('status', Unicode)
-    nummer = Column('nummer', Unicode)
-    x03 = Column('x03', Numeric)
-    y03 = Column('y03', Numeric)
-    n95 = Column('n95', Numeric)
-    e95 = Column('e95', Numeric)
-    h02 = Column('h02', Numeric)
-    zugang = Column('zugang', Unicode)
-    url = Column('url', Unicode)
-    bgdi_created = Column('bgdi_created', Unicode)
-    the_geom = Column(Geometry2D)
-
-register('ch.swisstopo.fixpunkte-lfp2', FixpunkteLfp2)
-
-
-class FixpunkteHfp2(Base, Vector):
-    __tablename__ = 'punkt_hoehe_hfp2'
-    __table_args__ = ({'schema': 'fpds', 'autoload': False})
-    __template__ = 'templates/htmlpopup/fixpunkte.mako'
-    __bodId__ = 'ch.swisstopo.fixpunkte-hfp2'
-    __label__ = 'id'
-    id = Column('pointid', Unicode, primary_key=True)
-    nbident = Column('nbident', Unicode)
-    punktname = Column('punktname', Unicode)
-    status = Column('status', Unicode)
-    nummer = Column('nummer', Unicode)
-    x03 = Column('x03', Numeric)
-    y03 = Column('y03', Numeric)
-    n95 = Column('n95', Numeric)
-    e95 = Column('e95', Numeric)
-    h02 = Column('h02', Numeric)
-    zugang = Column('zugang', Unicode)
-    url = Column('url', Unicode)
-    bgdi_created = Column('bgdi_created', Unicode)
-    the_geom = Column(Geometry2D)
-
-register('ch.swisstopo.fixpunkte-hfp2', FixpunkteHfp2)
