+++ chsdi/templates/htmlpopup/fixpunkte.mako
@@ -4,17 +4,14 @@
 <%
     nummer = c['layerBodId'] + '.' + 'id'
 %>
+
     <% c['stable_id'] = True %>
-	<tr><td class="cell-left">${_(nummer)}</td>         <td>${c['featureId']}</td></tr>
-<!--	
-    <tr><td class="cell-left">${_('name')}</td>           <td>${c['attributes']['punktname'] or '-'}</td></tr> 
--->
-	<tr><td class="cell-left">${_('status_fp')}</td>      <td>${c['attributes']['status'] or '-'}</td></tr>
-	<tr><td class="cell-left">${_('fp_Y03_X03')}</td>     <td>${c['attributes']['y03'] or '-'} / ${c['attributes']['x03'] or '-'}</td></tr>
-	<tr><td class="cell-left">${_('fp_E95_N95')}</td>     <td>${c['attributes']['e95'] or '-'} / ${c['attributes']['n95'] or '-'}</td></tr>
-	<tr><td class="cell-left">${_('fp_H02')}</td>         <td>${c['attributes']['h02'] or '-'}</td></tr>
-<!--	
-    <tr><td class="cell-left">${_('zugang')}</td>         <td>${c['attributes']['zugang'] or '-'}</td></tr> 
--->
-	<tr><td class="cell-left">${_('protokoll')}</td>      <td><a href="${c['attributes']['url'] or '-'}" target="_blank">${_('protokoll')}</a></td></tr>
+	<tr><td class="cell-left">${_(c['layerBodId']+'.id')}</td>						<td>${c['featureId']}</td></tr>
+	<tr><td class="cell-left">${_(c['layerBodId']+'.status')}</td>      			<td>${c['attributes']['status'] or '-'}</td></tr>
+	<tr><td class="cell-left">${_(c['layerBodId']+'.koordinate')}</td>      		<td>${c['attributes']['koordinate'] or '-'}</td></tr>
+	<tr><td class="cell-left">${_(c['layerBodId']+'.hoehe_geom_m')}</td>      		<td>${c['attributes']['hoehe_geom_m'] or '-'}</td></tr>
+	<tr><td class="cell-left">${_(c['layerBodId']+'.url_punktprotokoll')}</td>      <td><a href="${c['attributes']['url_punktprotokoll'] or '-'}" target="_blank">${_(c['layerBodId']+'.url_punktprotokoll')}</a></td></tr>
+	<tr><td class="cell-left">${_(c['layerBodId']+'.punktzeichen')}</td>      		<td>${c['attributes']['punktzeichen'] or '-'}</td></tr>
+	<tr><td class="cell-left">${_(c['layerBodId']+'.kanton')}</td>      			<td>${c['attributes']['kanton'] or '-'}</td></tr>
+	<tr><td class="cell-left">${_(c['layerBodId']+'.zust_name')}</td>      			<td>${c['attributes']['zust_name'] or '-'}</td></tr>
 </%def>
